+++ src/main/java/data/ScoringStrategy.java
@@ -3,5 +3,5 @@ package data;
 import java.util.Set;
 
 public interface ScoringStrategy {
-    int evaluate(Set<Integer> userAnswers, Set<Integer> correctAnswers);
+    double evaluate(Set<Integer> userAnswers, Set<Integer> correctAnswers);
 }
